pyproject.toml
[project]
name = "ws.webpasswd"
description = "Change a UNIX password via web UI"

dependencies = [
  "flask",
  "flask_wtf",
  "python-pam >= 2.0.0",
  "six",  # undeclared by python-pam
  "wtforms",
  "pyyaml",
]
optional-dependencies = {test=[
  "pytest",
  "pytest-cov",
]}

authors = [{name="Wolfgang Schnerring", email="wosc@wosc.de"}]
license = {text=" BSD-3-Clause"}
urls = {Repository="https://github.com/wosc/webpasswd"}
requires-python = ">=3.7"

dynamic = ["version", "readme"]

[project.scripts]
webpasswd-change = "ws.webpasswd.update:main"
webpasswd-cgi = "ws.webpasswd.web:cgi"
webpasswd-serve = "ws.webpasswd.web:serve"


[build-system]
requires = ["hatchling", "hatch-fancy-pypi-readme"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ws"]  # required for pep420 namespace package, pypa/hatch#587
strict-naming = false  # required for proper output filename
[tool.hatch.build.targets.sdist]
strict-naming = false  # required for proper output filename

[tool.hatch.version]
path = "src/ws/webpasswd/__init__.py"

[tool.hatch.metadata.hooks.fancy-pypi-readme]
content-type = "text/x-rst"
fragments = [{path="README.rst"}, {text="\n\n"}, {path="CHANGES.txt"}]


[tool.hatch.envs.default]
# Make work on older OS
python = "3"
dependencies = ["pip>=23"]


[tool.hatch.envs.test]
features = ["test"]

[tool.pytest.ini_options]
addopts = "-s --tb=native --cov=src --cov-report=html"


[tool.hatch.envs.release]
skip-install = true
dependencies = ["towncrier"]

[tool.towncrier]
directory = "changelog/"
filename = "CHANGES.txt"
title_format = ""  # changelog file already has a title
# First underline is used for version/date header.
# Second underline is used for the type names (like 'Bug fixes:').
underlines = ["-", "+"]

[[tool.towncrier.type]]
directory = "change"
name = "Changes"  # I'd love to omit the category, but that requires copy&paste of the whole template
showcontent = true
